Move scan related code to scan package
