+++ pir/pir-internal/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
             android:label="@string/pirDevViewScanResults" />
 
         <service
-            android:name=".service.PirForegroundScanService"
+            android:name=".scan.PirForegroundScanService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="specialUse"
             android:process=":pir" />
 
         <service
-            android:name=".service.PirRemoteWorkerService"
+            android:name=".scan.PirRemoteWorkerService"
             android:exported="false"
             android:permission="android.permission.BIND_JOB_SERVICE"
             android:process=":pir" />
